+++ crates/controller/src/cli/fix_acl.rs
@@ -12,7 +12,7 @@ use anyhow::{bail, Context, Error, Result};
 use clap::Parser;
 use controller_settings::Settings;
 use database::{Db, DbConnection};
-use db_storage::{events::Event, rooms::Room, users::User};
+use db_storage::{events::Event, module_resources::ModuleResource, rooms::Room, users::User};
 use kustos::prelude::*;
 use std::sync::Arc;
 
@@ -36,6 +36,10 @@ pub(super) struct Args {
     #[clap(long, default_value = "false")]
     skip_rooms: bool,
 
+    /// Skip fix of module resources permissions
+    #[clap(long, default_value = "false")]
+    skip_module_resources: bool,
+
     /// Skip fix of event permission fixes
     #[clap(long, default_value = "false")]
     skip_events: bool,
@@ -56,6 +60,7 @@ pub(super) async fn fix_acl(settings: Settings, args: Args) -> Result<()> {
             skip_users: false,
             skip_groups: false,
             skip_rooms: false,
+            skip_module_resources: false,
             skip_events: false,
         } => {
             // Only remove all policies if none of the skips are specified
@@ -81,6 +86,10 @@ pub(super) async fn fix_acl(settings: Settings, args: Args) -> Result<()> {
 
     if !args.skip_rooms {
         fix_rooms(&mut conn, &authz).await?;
+    }
+
+    if !args.skip_module_resources {
+        fix_module_resources(&mut conn, &authz).await?;
     }
 
     if !args.skip_events {
@@ -171,6 +180,36 @@ async fn fix_rooms(conn: &mut DbConnection, authz: &kustos::Authz) -> Result<()>
             .grant_user_access(user.id)
             .room_read_access(room.id)
             .room_write_access(room.id)
+            .finish();
+    }
+
+    authz.add_policies(policies).await?;
+
+    Ok(())
+}
+
+async fn fix_module_resources(conn: &mut DbConnection, authz: &kustos::Authz) -> Result<()> {
+    let module_resources_with_creator = ModuleResource::get_all_with_creator_and_owner(conn)
+        .await
+        .context("failed to load module resources")?;
+
+    let mut policies = PoliciesBuilder::new();
+
+    for (module_resource_id, creator_id, owner_id) in module_resources_with_creator {
+        policies = policies
+            .grant_user_access(creator_id)
+            .add_resource(
+                module_resource_id.resource_id(),
+                [AccessMethod::Get, AccessMethod::Put, AccessMethod::Delete],
+            )
+            .finish();
+
+        policies = policies
+            .grant_user_access(owner_id)
+            .add_resource(
+                module_resource_id.resource_id(),
+                [AccessMethod::Get, AccessMethod::Put, AccessMethod::Delete],
+            )
             .finish();
     }