+++ crates/types/src/core/module_resource_id.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+crate::diesel_newtype! {
+    feature_gated:
+
+    #[derive(Copy)] ModuleResourceId(uuid::Uuid) => diesel::sql_types::Uuid, "/module_resources/"
+}